pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "layer-rag-query"
version = "1.0.0"
description = "RAG hybrid retrieval: dense + BM25 + RRF fusion"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = []
keywords = ["rag", "retrieval", "qdrant", "bm25", "rrf"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "qdrant-client>=1.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.20.0",
    "rank-bm25>=0.2.2",
    "tb-loki-central-logger>=0.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
mcp = [
    "fastmcp>=2.0",
    "fastapi>=0.100",
    "uvicorn[standard]>=0.22",
]

[tool.setuptools.packages.find]
where = ["."]
